api/requirements.txt
flask
requests
pandas
sqlalchemy
sqlalchemy-utils
api-offres-emploi